+++ mmpm/api/endpoints/ep_configs.py
@@ -1,14 +1,12 @@
-from pathlib import PosixPath
 from shutil import copyfile
-from typing import Dict
 
 from flask import Blueprint, Response, request, send_file
 
 from mmpm.api.constants import http
 from mmpm.api.endpoints.endpoint import Endpoint
-from mmpm.constants import paths
 from mmpm.env import MMPMEnv
 from mmpm.log.factory import MMPMLogFactory
+from mmpm.magicmirror.magicmirror import MagicMirrorConfigs
 
 logger = MMPMLogFactory.get_logger(__name__)
 
@@ -24,12 +22,7 @@ class Configs(Endpoint):
         self.blueprint = Blueprint(self.name, __name__, url_prefix=f"/api/{self.name}")
         self.handler = None
         self.env = MMPMEnv()
-        self.files: Dict[str, PosixPath] = {
-            "mmpm-env.json": paths.MMPM_ENV_FILE,
-            "config.js": self.env.MMPM_MAGICMIRROR_ROOT.get() / "config" / "config.js",
-            "custom.css": self.env.MMPM_MAGICMIRROR_ROOT.get() / "css" / "custom.css",
-            "config.js.sample": self.env.MMPM_MAGICMIRROR_ROOT.get() / "config" / "config.js.sample",
-        }
+        self.mm_configs = MagicMirrorConfigs()
 
         @self.blueprint.route("/retrieve/<filename>", methods=[http.GET])
         def retrieve_mmpm_env_json(filename: str) -> Response:
@@ -42,17 +35,20 @@ class Configs(Endpoint):
             Returns:
                 Response: A Flask Response object containing the requested file or an error message.
             """
-            if filename not in self.files:
-                return self.failure(f"File '{filename}' is not recognized. Only {list(self.files.keys())} are valid.")
+            try:
+                file = self.mm_configs.get(filename)
+            except (KeyError, Exception):
+                return self.failure(f"File '{filename}' is not recognized")
 
-            file = self.files.get(filename)
-
-            if filename == "config.js" and (not file.exists() or not bool(file.stat().st_size)):
-                sample = self.files.get("config.js.sample")
-
-                if sample.exists():
-                    copyfile(sample, file)
+            if filename == self.mm_configs.config_js.name and (not file.exists() or not bool(file.stat().st_size)):
+                if self.mm_configs.config_js_sample.exists():
+                    copyfile(self.mm_configs.config_js_sample, file)
                     logger.debug("The config.js hasn't been initialized yet, but found the sample file. Copying config.js.sample to config.js")
+
+            elif filename == self.mm_configs.custom_css.name and (not file.exists() or not bool(file.stat().st_size)):
+                if self.mm_configs.custom_css_sample.exists():
+                    copyfile(self.mm_configs.custom_css_sample, file)
+                    logger.debug("The custom.css hasn't been initialized yet, but found the sample file. Copying custom.css.sample to custom.css")
 
             if not file.exists():
                 logger.debug(f"File '{file}' does not exist, creating empty file")
@@ -61,7 +57,7 @@ class Configs(Endpoint):
 
             logger.debug(f"Sending back {file}")
 
-            return send_file(self.files.get(filename), filename, as_attachment=True)
+            return send_file(file, filename, as_attachment=True)
 
         @self.blueprint.route("/update/<filename>", methods=[http.POST])
         def update_mmpm_env_json(filename: str) -> Response:
@@ -75,10 +71,11 @@ class Configs(Endpoint):
                 Response: A Flask Response object indicating the success or failure of the update operation.
             """
 
-            if filename not in self.files:
-                return self.failure(f"File '{filename}' is not recognized. Only {list(self.files.keys())} are valid.")
+            try:
+                file = self.mm_configs.get(filename)
+            except (KeyError, Exception):
+                return self.failure(f"File '{filename}' is not recognized")
 
-            file = self.files.get(filename)
             logger.debug(f"Editing {file} with provided contents")
 
             contents = request.get_json().get("contents")